+++ mono_repo/lib/src/mono_config.g.dart
@@ -1,0 +1,15 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'mono_config.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+MonoConfig _$MonoConfigFromJson(Map json) {
+  return $checkedNew('MonoConfig', json, () {
+    var val = MonoConfig($checkedConvert(json, 'travis',
+        (v) => (v as Map)?.map((k, e) => MapEntry(k as String, e))));
+    return val;
+  });
+}